refactor[#231] : mapSnackbar.tsx 추가
