+++ src/entities/map/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./customMarker";
 export * from "./currentLocationLoading";
+export * from "./mapSnackbar";
